fix: move upload-drain refresh to Dropzone so it fires in the file view
The refresh-on-drain effect was in DirectoryList, which the page renders
only for directory views — when the path resolves to a single file, page.tsx
returns the ObjectSummary/ObjectPreview branch and never mounts DirectoryList.
So re-uploading while viewing that exact file never triggered a refresh.

Move the effect into Dropzone, which lives in the product layout and is
mounted for both directory and file views. router.refresh() now re-runs the
server render (listing, file metadata, and server-fetched internal previews)
after either kind of re-upload.
